+++ src/classes/babies/NoArmsBaby.ts
@@ -3,7 +3,7 @@ import {
   ModCallback,
   PickupPrice,
 } from "isaac-typescript-definitions";
-import { Callback } from "isaacscript-common";
+import { Callback, asNumber } from "isaacscript-common";
 import { PICKUP_VARIANTS_IMMUNE_TO_BABY_EFFECTS } from "../../constants";
 import { Baby } from "../Baby";
 
@@ -15,7 +15,7 @@ export class NoArmsBaby extends Baby {
 
     if (
       !PICKUP_VARIANTS_IMMUNE_TO_BABY_EFFECTS.has(pickup.Variant) &&
-      pickup.Price === (PickupPrice.NULL as int) // We don't want it to affect shop items.
+      pickup.Price === asNumber(PickupPrice.NULL) // We don't want it to affect shop items.
     ) {
       // Make it impossible for the player to pick up this pickup.
       if (pickup.EntityCollisionClass !== EntityCollisionClass.NONE) {