+++ src/classes/babies/RictusBaby.ts
@@ -3,7 +3,7 @@ import {
   ModCallback,
   PickupPrice,
 } from "isaac-typescript-definitions";
-import { Callback, isHeart } from "isaacscript-common";
+import { Callback, asNumber, isHeart } from "isaacscript-common";
 import { PICKUP_VARIANTS_IMMUNE_TO_BABY_EFFECTS } from "../../constants";
 import { Baby } from "../Baby";
 
@@ -16,7 +16,7 @@ export class RictusBaby extends Baby {
     if (
       !PICKUP_VARIANTS_IMMUNE_TO_BABY_EFFECTS.has(pickup.Variant) &&
       !isScaredHeart(pickup) &&
-      pickup.Price === (PickupPrice.NULL as int) && // We don't want it to affect shop items.
+      pickup.Price === asNumber(PickupPrice.NULL) && // We don't want it to affect shop items.
       pickup.Position.Distance(player.Position) <= 80
     ) {
       pickup.Velocity = pickup.Position.sub(player.Position)