+++ src/constants.ts
@@ -222,3 +222,9 @@ export const COLLECTIBLES_THAT_REMOVE_TEARS = [
   CollectibleType.SPIRIT_SWORD, // 579
   CollectibleType.BERSERK, // 704
 ] as const;
+
+export const BLINDFOLDED_ANTI_SYNERGY_COLLECTIBLES = [
+  CollectibleType.MAW_OF_THE_VOID, //399
+  CollectibleType.REVELATION, //643
+  CollectibleType.MONTEZUMAS_REVENGE, //680
+] as const;