+++ docs/history.md
@@ -3,7 +3,7 @@
 <!-- cspell:ignore Adrayon,cantgirl,Fallenfor,Finalkids,Hispa,Kaethela,lenamek,mgln,Moucheron,Quipet,Sanio,sisuka,SuperpaperChildy,Toooschi -->
 <!-- markdownlint-disable MD001 -->
 
-### _v1.9.1_ - September 11th, 2023
+### _v1.10.0_ - September 11th, 2023
 
 - Changed the following babies:
   - Bone Baby (#57) - Reveals a random room on room clear